+++ packages/thumbnail/src/ThumbnailList.tsx
@@ -25,21 +25,25 @@ import type { RenderThumbnailItem } from './types/RenderThumbnailItemProps';
 export const ThumbnailList: React.FC<{
     currentPage: number;
     doc: PdfJs.PdfDocument;
+    pagesRotation: Map<number, number>;
     pageHeight: number;
     pageWidth: number;
     renderCurrentPageLabel?: RenderCurrentPageLabel;
     renderThumbnailItem?: RenderThumbnailItem;
     rotation: number;
     onJumpToPage(pageIndex: number): void;
+    onRotatePage(pageIndex: number, rotation: number): void;
 }> = ({
     currentPage,
     doc,
+    pagesRotation,
     pageHeight,
     pageWidth,
     renderCurrentPageLabel,
     renderThumbnailItem,
     rotation,
     onJumpToPage,
+    onRotatePage,
 }) => {
     const [labels, setLabels] = React.useState([]);
     const { numPages } = doc;
@@ -216,11 +220,14 @@ export const ThumbnailList: React.FC<{
                     ? renderCurrentPageLabel({ currentPage, pageIndex, numPages, pageLabel })
                     : pageLabel;
 
+                const pageRotation = pagesRotation.has(pageIndex) ? pagesRotation.get(pageIndex) : 0;
+
                 const thumbnail = (
                     <ThumbnailContainer
                         doc={doc}
                         pageHeight={pageHeight}
                         pageIndex={pageIndex}
+                        pageRotation={pageRotation}
                         pageWidth={pageWidth}
                         rotation={rotation}
                         shouldRender={renderPageIndex === pageIndex}
@@ -238,6 +245,7 @@ export const ThumbnailList: React.FC<{
                         renderPageLabel: <>{label}</>,
                         renderPageThumbnail: thumbnail,
                         onJumpToPage: () => onJumpToPage(pageIndex),
+                        onRotatePage: (rotation: number) => onRotatePage(pageIndex, rotation),
                     })
                 ) : (
                     <div key={key}>