+++ packages/thumbnail/src/thumbnailPlugin.tsx
@@ -35,7 +35,15 @@ export interface ThumbnailPluginProps {
 }
 
 export const thumbnailPlugin = (pluginProps?: ThumbnailPluginProps): ThumbnailPlugin => {
-    const store = React.useMemo(() => createStore<StoreProps>({}), []);
+    const store = React.useMemo(
+        () =>
+            createStore<StoreProps>({
+                rotatePage: () => {
+                    /**/
+                },
+            }),
+        []
+    );
     const [docId, setDocId] = React.useState('');
 
     const CoverDecorator = (props: CoverProps) => (
@@ -58,6 +66,7 @@ export const thumbnailPlugin = (pluginProps?: ThumbnailPluginProps): ThumbnailPl
     return {
         install: (pluginFunctions: PluginFunctions) => {
             store.update('jumpToPage', pluginFunctions.jumpToPage);
+            store.update('rotatePage', pluginFunctions.rotatePage);
         },
         onDocumentLoad: (props: PluginOnDocumentLoad) => {
             setDocId(props.doc.loadingTask.docId);
@@ -65,6 +74,7 @@ export const thumbnailPlugin = (pluginProps?: ThumbnailPluginProps): ThumbnailPl
         },
         onViewerStateChange: (viewerState: ViewerState) => {
             store.update('currentPage', viewerState.pageIndex);
+            store.update('pagesRotation', viewerState.pagesRotation);
             store.update('pageHeight', viewerState.pageHeight);
             store.update('pageWidth', viewerState.pageWidth);
             store.update('rotation', viewerState.rotation);