+++ packages/thumbnail/src/types/StoreProps.ts
@@ -11,8 +11,10 @@ import type { PdfJs } from '@react-pdf-viewer/core';
 export interface StoreProps {
     currentPage?: number;
     doc?: PdfJs.PdfDocument;
+    pagesRotation?: Map<number, number>;
     pageHeight?: number;
     pageWidth?: number;
     jumpToPage?(pageIndex: number): void;
+    rotatePage(pageIndex: number, rotation: number): void;
     rotation?: number;
 }