+++ CHANGELOG.md
@@ -16,6 +16,46 @@
 />
 ```
 
+-   Provide the ability of rotating a particular page. You can customize a thumbnail renderer to add the rotating functionality to each page:
+
+```tsx
+const renderThumbnailItem = (props: RenderThumbnailItemProps) => (
+    <MinimalButton onClick={() => props.onRotatePage(90)}>
+        <RotateForwardIcon />
+    </MinimalButton>
+    <MinimalButton onClick={() => props.onRotatePage(-90)}>
+        <RotateBackwardIcon />
+    </MinimalButton>
+);
+
+const thumbnailPluginInstance = thumbnailPlugin();
+const { Thumbnails } = thumbnailPluginInstance;
+
+<Thumbnails renderThumbnailItem={renderThumbnailItem} />
+```
+
+It's also possible to do it by using the `renderPage`:
+
+```tsx
+const renderPage: RenderPage = (props: RenderPageProps) => (
+    <>
+        {props.canvasLayer.children}
+        <div>
+            <MinimalButton onClick={() => props.rotatePage(90)}>
+                <RotateForwardIcon />
+            </MinimalButton>
+            <MinimalButton onClick={() => props.rotatePage(-90)}>
+                <RotateBackwardIcon />
+            </MinimalButton>
+        </div>
+        {props.annotationLayer.children}
+        {props.textLayer.children}
+    </>
+);
+
+<Viewer renderPage={renderPage} />;
+```
+
 **Improvements**
 
 -   The search popover is opened if users press the shortcuts (`Ctrl + F`, or `Cmd + F` on macOS) when the mouse is inside the viewer container